Split AccountDetailsDropdown into container and component files
