+++ ui/app/components/app/dropdowns/account-details-dropdown/index.js
@@ -1,0 +1,1 @@
+export { default } from './account-details-dropdown.container'
